server/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
python-dotenv
tweepy
